requirements.txt
grpcio==1.33.2
h5py==2.10.0
Keras-Preprocessing==1.1.2
mypy-extensions==0.4.3
numpy==1.18.5
pandas==1.1.4
proto-plus==1.11.0
protobuf==3.13.0
requests==2.25.0
requests-oauthlib==1.3.0
scikit-learn==0.23.1
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-metadata==0.25.0
tensorflow-serving-api==2.3.0
tensorflow-text==2.3.0
tensorflow-transform==0.25.0
